pyproject.toml
[project]
name = "ai-product-photo-detector"
version = "0.1.0"
description = "Detect AI-generated product photos in e-commerce listings"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Nolan Cacheux", email = "cachnolan@gmail.com" }
]
keywords = ["mlops", "ai-detection", "computer-vision", "pytorch", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # ML & Deep Learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "pillow>=10.0.0",
    
    # API
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
    "pyjwt>=2.8.0",
    
    # MLOps
    "mlflow>=2.8.0",
    "dvc>=3.30.0",
    "optuna>=3.4.0",
    
    # Observability
    "structlog>=23.1.0",
    "prometheus-client>=0.17.0",
    
    # Data & Config
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Utils
    "httpx>=0.24.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

ui = [
    "streamlit>=1.28.0",
]

[project.urls]
Homepage = "https://github.com/nolancacheux/mlops_project"
Repository = "https://github.com/nolancacheux/mlops_project"
Documentation = "https://github.com/nolancacheux/mlops_project/tree/main/docs"

[project.scripts]
train = "src.training.train:main"
serve = "src.inference.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
